Proper macOS libLLVM symlink when cross compiling
When cross compiling on macOS with `llvm.link-shared` enabled,
the symlink creation will fail after compiling LLVM for the target
architecture, because it will attempt to create the symlink in the
host LLVM directory, which was already created when being built.

This commit changes the symlink path to the actual LLVM output.