respect CARGO_PKG_RUST_VERSION when issuing deprecation warnings
you can specify the `rust-version`[0] in `Cargo.toml` file.
If you specify that, we should not emit a deprecated warning.
Reason for this is, that code will be sprinkled with
`#[allow(deprecated)]` and maybe be forgotten in the future, so
deprecated items will be used.
If you specify the `rust-version` field, no such deprecation warning
should be emitted, unless you upgrade your msrv.

[0]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field