Be stricter about combining `repr(packed)` with `derive`.
Currently there are two errors you can get when using `derive` on a
`repr(packed)` struct:

- "`#[derive]` can't be used on a `#[repr(packed)]` struct with type
  or const parameters (error E0133)"
- "`#[derive]` can't be used on a `#[repr(packed)]` struct that does not
  derive Copy (error E0133)"

These were introduced under #82523, because they can lead to unaligned
references, which is UB. They are in the process of being upgraded to
hard errors.

This is all fine, but the second error overstates things. It *is*
possible to use `derive` on a `repr(packed)` struct that doesn't derive
`Copy` in two cases.
- If all the fields within the struct meet the required alignment: 1 for
  `repr(packed)`, or N for `repr(packed(N))`.
- Or, If `Default` is the only trait derived.

These exceptions don't seem particularly important or useful, and the
language would be simpler if they are removed. So this commit does that,
making these cases a hard error. The new checks are done at derive code
creation time which means that `unsafe_derive_on_repr_packed` is no
longer needed at MIR build time.

Test changes:
- deriving-with-repr-packed.rs has numerous changes to the code to
  broaden coverage. Some of the errors are reported twice now,
  which is hard to avoid but harmless (and users won't see duplicates
  thanks to the normal error de-duplication). Also, if a packed struct
  is both non-`Copy` and has type params, both those errors are now
  reported.
- deriving-all-codegen.rs has the `PackedNonCopy` case removed, because
  it's no longer allowed.
- packed.rs needs `Copy` added to some packed struct that currently only
  derive `Default`.