Loosen shadowing check inside macro contexts.
This allows macro code to use identifiers for local variables which
are also defined as statics/consts/ctors outside of the macro. This
improves robustness of macro hygiene, as identifiers used inside a
macro can no longer produce unexpected restrictions on names for
statics and consts defined outside of it. Concretely, constructions
such as
```rust
thread_local!(static FOO: usize = 0);
static init: usize = 0;
```
no longer unexpectedly fail to compile because thread_local! happens
to define a function with init as a parameter name.