[perf] Don't emit noalias for box when compiling rustc itself
This compiles rustc without noalias on box to evaluate the performance
impact of such a change on a large scale project like rustc. Most
importantly, this does NOT change the emitting of noalias when compiling
crates other than rustc, to avoid perf impacts from LLVM having to
process less attributes like in #98017.