Explain how Vec::with_capacity is faithful
There are concerns that the doc changes in rust-lang/rust@95dc353006c23a4493b3d08ba33680e51c2107c8
are breaking changes in the promised API. In this commit, I explain in more detail
the exact promise that Vec::with_capacity is really making:
Vec is trying to act unsurprising by relaying information faithfully to the allocator,
it is not committing to internal details of Vec itself beyond that.
As it happens, we don't get useful capacity information from allocators,
but once upon a time Rust did capacity recalculation from available data,
and we should reserve the right to do so again if it seems profitable and correct.

This path avoids adding a duplicate `with_capacity_exact` to Vec's already-formidable API surface.