Allow search field focus with slash (/) in rustdoc web
This patch adds support for recognizing `/` as a "jump to search field"
shortcut, akin to `s` and `S`. This new shortcut isn't (yet?) advertised
anywhere, it's merely supported as an "alias" for the current shortcut keys.

Rationale: search in many TUIs has traditionally been mapped to `/` for
compatibility with VIM. A number of websites/apps also respect this convention
(most notably, google.com). (It's also way nerdier than `s` and that should be
reason enough to support it!)