Add invalid UTF-8 example to `BufRead::read_line()`
Add some text and an associated example to clarify the behavior of
`BufRead::read_line()` when it comes to errors arising from the attempted
conversion of invalid UTF-8 sequences. There really wasn't much documentation on
how UTF-8 errors can be handled when calling `.lines()` or `.read_line()`
without digging into the source. Without this, people might naively expect that
the same error as `str::from_utf8()` would be emitted (which isn't the case).