DO NOT MERGE - scratchwork
Punch a hole to allow direct access to bytes - this API should be
changed to provide a `Size` value corresponding to an `AllocId`.

Additionally, force bytes to be appropriately aligned - the current
trick is a hack, and over-aligns. We should use the `Align` parameter to
appropriately align these buffers, as otherwise they'll never be able to
be used by native code.

The "adjust" callback used on deserialized allocations might not have
that information - if it doesn't, just overalign to max alignment