Check that feature stabilizations contain the current version
It's a common phenomenon that feature stabilizations don't make it into
a particular release, but the version is still inaccurate. Often this
leads to subsequent changes to adjust/correct the version, but it's easy
to miss errors.

This commit adds checking to ensure that a stabilization is always
for the current release. Adjusting versions of already stable features
is still supported, as is unstabilizing or other modifications.